+++ packages/api/src/modules/notify/outPipes/BrevoContactNotifyPipe.test.ts
@@ -2,24 +2,33 @@ import {
     AdminTerritorialLevel,
     AgentJobTypeEnum,
     AgentTypeEnum,
-    RegistrationSrcTypeEnum,
     TerritorialScopeEnum,
+    RegistrationSrcTypeEnum,
 } from "dto";
-import { API_SENDINBLUE_CONTACT_LIST } from "../../../configurations/apis.conf";
+import { API_BREVO_CONTACT_LIST } from "../../../configurations/apis.conf";
 import { NotificationType } from "../@types/NotificationType";
-import BrevoContactNotifyPipe from "./BrevoContactNotifyPipe";
+import BrevoContactNotifyPipe, {
+    BrevoContactNotifyPipe as BrevoContactNotifyPipeClass,
+} from "./BrevoContactNotifyPipe";
+import * as Brevo from "@getbrevo/brevo";
+jest.mock("@getbrevo/brevo");
 
 describe("BrevoContactNotifyPipe", () => {
     const USER_EMAIL = "user@beta.gouv.fr";
 
+    const mockRequestContactImport = jest.spyOn(Brevo, "RequestContactImport");
     // @ts-expect-error apiInstance is private attribute
     const mockUpdateContact = jest.spyOn(BrevoContactNotifyPipe.apiInstance, "updateContact");
     // @ts-expect-error apiInstance is private attribute
     const mockDeleteContact = jest.spyOn(BrevoContactNotifyPipe.apiInstance, "deleteContact");
+    // @ts-expect-error apiInstance is private attribute
+    const mockImportContacts = jest.spyOn(BrevoContactNotifyPipe.apiInstance, "importContacts");
 
     beforeAll(() => {
         mockUpdateContact.mockResolvedValue({ body: { id: 1 } } as any);
         mockDeleteContact.mockResolvedValue({ body: { id: 1 } } as any);
+        mockImportContacts.mockResolvedValue({ body: { id: 1 } } as any);
+        mockRequestContactImport.mockImplementation(() => ({}));
     });
 
     describe("notify", () => {
@@ -70,7 +79,7 @@ describe("BrevoContactNotifyPipe", () => {
                     SOURCE_IMPORT: "Data.Subvention",
                     LIEN_ACTIVATION: "TOKEN",
                 },
-                listIds: [Number(API_SENDINBLUE_CONTACT_LIST)],
+                listIds: [Number(API_BREVO_CONTACT_LIST)],
             };
 
             // @ts-expect-error userCreated is private method
@@ -91,7 +100,7 @@ describe("BrevoContactNotifyPipe", () => {
             await BrevoContactNotifyPipe.userActivated({ email: USER_EMAIL });
             expect(mockUpdateContact).toHaveBeenCalledWith(USER_EMAIL, {
                 attributes: { COMPTE_ACTIVE: true, LIEN_ACTIVATION: "" },
-                listIds: [Number(API_SENDINBLUE_CONTACT_LIST)],
+                listIds: [Number(API_BREVO_CONTACT_LIST)],
             });
         });
     });
@@ -106,7 +115,7 @@ describe("BrevoContactNotifyPipe", () => {
 
             expect(mockUpdateContact).toHaveBeenCalledWith(email, {
                 attributes: { DERNIERE_CONNEXION: now },
-                listIds: [Number(API_SENDINBLUE_CONTACT_LIST)],
+                listIds: [Number(API_BREVO_CONTACT_LIST)],
             });
         });
     });
@@ -169,6 +178,25 @@ describe("BrevoContactNotifyPipe", () => {
             //@ts-expect-error: private method
             await BrevoContactNotifyPipe.batchUsersDeleted(UPDATE_PAYLOAD);
             expect(mockDeleteContact.mock.calls).toMatchSnapshot();
+        });
+    });
+
+    describe("updateNbRequests", () => {
+        it("should create a Brevo RequestContactImport", () => {
+            // @ts-expect-error: test private method
+            BrevoContactNotifyPipe.updateNbRequests([]);
+            expect(jest.mocked(Brevo.RequestContactImport)).toHaveBeenCalledTimes(1);
+        });
+
+        it("should call importContacts", () => {
+            const data = [
+                { email: "john.doe@gouv.fr", nbVisits: 13 },
+                { email: "arsene.lupin@gouv.fr", nbVisits: "45" },
+            ];
+            // @ts-expect-error: test private method
+            BrevoContactNotifyPipe.updateNbRequests(data);
+            const expected = data.map(d => ({ email: d.email, attributes: { NB_REQUETES: d.nbVisits } }));
+            expect(mockImportContacts).toHaveBeenCalledWith({ jsonBody: expected, listIds: [1] });
         });
     });
 });