+++ packages/front/src/routes/dataviz/DataViz.controller.ts
@@ -1,0 +1,10 @@
+import type { AmountsVsProgramRegionDto } from "dto";
+import dataVizService from "$lib/resources/dataViz/dataViz.service";
+
+export class DataVizController {
+    amountsVsProgramRegionDataPromise: Promise<AmountsVsProgramRegionDto[]>;
+
+    constructor() {
+        this.amountsVsProgramRegionDataPromise = dataVizService.getAmountsVsProgramRegion();
+    }
+}